requirements.txt
python-dotenv>=1.0.0       
pydantic>=2.0.0             
openai>=1.30.0
regex>=2024.0.0
sqlalchemy>=2.0.0
dbt-postgres>=1.7.0
pandas>=2.0.0               
pytest>=8.0.0
fastapi>=0.100.0
uvicorn>=0.30.0
pymssql>=2.2.0